Move JavaScript tests to component folder
- Makes components more portable. For example when we need to move components from this repo to CMS or Framework there's just one directory to move.
