+++ javascript/src/state/gallery/action-types.js
