+++ javascript/src/state/gallery/reducer.js
@@ -1,5 +1,5 @@
 import deepFreeze from 'deep-freeze';
-import { GALLERY } from '../action-types';
+import { GALLERY } from './action-types';
 import CONSTANTS from '../../constants.js';
 
 const initialState = {
