+++ package.json
@@ -50,7 +50,6 @@
   },
   "jest": {
     "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
-    "testDirectoryName": "tests/react",
     "unmockedModulePathPatterns": [
       "<rootDir>/node_modules/react"
     ],
